+++ proptest/test_axioms.py
@@ -21,6 +21,8 @@ from discopy import (
     symmetric,
     traced,
 )
+from discopy import tensor
+from discopy.frobenius import Dim
 from discopy.matrix import Matrix
 from discopy.testing import assert_verdict
 from discopy.utils import factory_name
@@ -48,6 +50,7 @@ CARRIERS = (
     frobenius.Ty, frobenius.Diagram, frobenius.Hypergraph, frobenius.CMap,
     frobenius.Functor,
     Matrix[int],
+    Dim, tensor.Diagram, tensor.Tensor[int],
 )